feat: Implement two-screen homepage with 404 Not Found page
- Create 404 Not Found page at root with Google-inspired design
- Display error message in Japanese (ページが見つかりません)
- Add back button to return to previous page
- Add link to home page (/home)
- Move original homepage to /home route
- Add animated transitions using Motion for visual polish

The top page now displays a 404 Not Found experience styled after Google's 404 page, with a "Butter" logo (using Google colors). Users can navigate back using the "戻る" (back) button or access the main homepage using the "ホームページへ" (go to homepage) link.
